JEL-299: Adapt to message tooling rework
