+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/Messenger/ProductsWereUpdatedHandler.php
@@ -10,16 +10,12 @@ use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ProductUuid;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ProductUuidCollection;
 use Akeneo\Pim\Enrichment\Product\API\Event\ProductsWereUpdated;
 use Akeneo\Pim\Enrichment\Product\API\Event\ProductWasUpdated;
-use Akeneo\Tool\Component\Messenger\PublicEventMessageInterface;
-use Akeneo\Tool\Component\Messenger\TraceableMessageHandlerInterface;
-use Akeneo\Tool\Component\Messenger\TraceableMessageInterface;
-use Webmozart\Assert\Assert;
 
 /**
  * @copyright 2023 Akeneo SAS (https://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-final class ProductsWereUpdatedHandler implements TraceableMessageHandlerInterface
+final class ProductsWereUpdatedHandler
 {
     public function __construct(
         private readonly CreateCriteriaEvaluations $createProductCriteriaEvaluations,
@@ -27,10 +23,8 @@ final class ProductsWereUpdatedHandler implements TraceableMessageHandlerInterfa
     ) {
     }
 
-    public function __invoke(TraceableMessageInterface $message): void
+    public function __invoke(ProductsWereUpdated $message): void
     {
-        Assert::isInstanceOf($message, ProductsWereUpdated::class);
-
         $productUuids = [];
         /** @var ProductWasUpdated $productWasUpdated */
         foreach ($message->events as $productWasUpdated) {
